publish site to docs folder instead of root
We currently publish the site to the root of the gh-pages
branch, but this is annoying since it means *everything* in
that branch is published. If we want to have some additional
stuff in the branch that isn't published (like github workflow
files, or screenshots), we need to publish to a subdirectory
instead.

The only option is /docs because of a limtiation of
github pages.