Site build for: 39a0ee0bb79dac628fe6f7ee96dda3eaa01a609f make the dark mode javascript resilient
localStorage may not be available on old browsers or if the user has
disabled it, but we can't leave an unstyled page in that case!

So we fall back to the media query if LS fails, and also a final fallback
to light mode if there is an unexpected exception.