Site build for: eec1362f7ee4ed8e5d53c0d1894bfae7fe0241d5 make the dark mode javascript resilient
localStorage may not be available on old browsers or if the user has
disabled it, but we can't leave an unstyled page in that case!

So we fall back to the media query if LS fails, and also a final fallback
to light mode if there is an unexpected exception.